requirements.txt
# --- Core app ---
streamlit==1.48.0
numpy==2.3.2
pandas==2.3.2
python-dotenv==1.0.1
altair==5.3.0

# --- DB / SQL ---
SQLAlchemy==2.0.32
psycopg[binary]==3.2.9

# --- LangChain stack (NumPy 2–friendly) ---
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.27
langchain-openai==0.3.31
openai==1.101.0

# --- Vector store ---
chromadb==1.0.20

# --- Extras ---
pydantic==2.8.2
typing_extensions==4.12.2
tiktoken==0.11.0
httpx==0.27.0
psycopg2
faiss-cpu
openpyxl
